Fixed the philopatry check procedure
philopatry-check was favouring the islands that came first numerically and was incorrectly allocating birds. This has been fixed and no particular island is favoured by number. However, all curves currently lead to one island siphoning off all birds as the probability tends towards 1. This might be fixed with an overhaul of the emigration procedure which currently takes birds that have failed to breed several times and removes them from the system. Instead, this could be rewritten to allow the birds to have the chance to establish on one of the islands in the system, that is not currently the most densely occupied (i.e. remove the highest probability island and rescale attractiveness among the remaining islands.)