find-mate fix in the emigration transition
Updated a line in the withinSeason script for the emigration pool. Individuals did not have their burrow wiped which meant that they could technically return to their own burrow problematically. Hopefully, this fixes the issue where very infrequently an individual cannot execute the find-mate. Not sure if it has though!